Refactor proxy operators into separate module
